requirements.txt
fastapi
uvicorn
pyyaml
streamlit
